requirements.txt
altair==5.1.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
entrypoints==0.4
extra-streamlit-components==0.1.60
Faker==19.12.0
favicon==0.7.0
firebase-admin==6.2.0
fonttools==4.43.1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.0
grpcio-status==1.59.0
htbuilder==0.6.2
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
msgpack==1.0.7
numpy==1.26.1
packaging==23.2
pandas==2.1.1
password-strength==0.0.3.post2
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
pymdown-extensions==10.3.1
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.26.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-image-select==0.6.0
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
zipp==3.17.0
